HuffmanCode

+++ MSDB/MSDB-CORE/include/compression/huffmanChunk.h
@@ -69,7 +69,7 @@ private:
 		//	++(*bit);
 		//}
 
-		aHuffmanCoder<uint32_t, uint8_t> coder(sizeof(uint8_t) * CHAR_BIT);
+		aHuffmanCoder<uint16_t, uint8_t> coder(sizeof(uint8_t) * CHAR_BIT);
 		auto buffer = (unsigned char*)this->getBuffer()->getData();
 		auto size = this->getBuffer()->size();
 		for(int i = 0; i < size; ++i)
